pull const_eval into separate module
