+++ flux-driver/src/collector.rs
@@ -5,6 +5,7 @@ use flux_common::{
     iter::IterExt,
 };
 use flux_errors::{FluxSession, ResultExt};
+use flux_middle::rty::Constant;
 use flux_syntax::{
     parse_expr, parse_flux_item, parse_fn_surface_sig, parse_qual_names, parse_refined_by,
     parse_ty, parse_type_alias, parse_variant, surface, ParseResult,
@@ -60,7 +61,7 @@ pub(crate) struct FnSpec {
 #[derive(Debug)]
 pub(crate) struct ConstSig {
     pub _ty: surface::ConstSig,
-    pub val: i128,
+    pub val: Constant,
 }
 
 macro_rules! attr_name {
@@ -150,9 +151,16 @@ impl<'tcx, 'a> SpecCollector<'tcx, 'a> {
             return Err(self.emit_err(errors::InvalidConstant { span }))
         };
 
+        // 0x8000000000000000
+        // 16 * 4 = 64
         let size = val.size();
-        if let Ok(val) = val.try_to_int(size) {
-            self.specs.consts.insert(def_id, ConstSig { _ty, val });
+        let conv = val.try_to_u64();
+        println!("TRACE: parse_const_spec: {def_id:?} ({size:?}) is {conv:?} is {val:?}");
+        if let Ok(val) = conv {
+            println!("TRACE: parse_const_spec: {def_id:?} = {val:?}");
+            self.specs
+                .consts
+                .insert(def_id, ConstSig { _ty, val: val as i128 });
             Ok(())
         } else {
             Err(self.emit_err(errors::InvalidConstant { span }))
